feat: data source table system field
