+++ frontend/packages/web/src/views/customer/components/openSeaTable.vue
@@ -619,17 +619,20 @@
     openSea.value = openSeaOptions.value[0]?.id || '';
   }
 
-  const filterConfigList = computed(() => [
-    {
-      title: t('customer.recycleReason'),
-      dataIndex: 'reasonId',
-      type: FieldTypeEnum.SELECT_MULTIPLE,
-      selectProps: {
-        options: reasonOptions.value,
-      },
-    },
-    ...baseFilterConfigList,
-  ]);
+  const filterConfigList = computed(
+    () =>
+      [
+        {
+          title: t('customer.recycleReason'),
+          dataIndex: 'reasonId',
+          type: FieldTypeEnum.SELECT_MULTIPLE,
+          selectProps: {
+            options: reasonOptions.value,
+          },
+        },
+        ...baseFilterConfigList,
+      ] as FilterFormItem[]
+  );
 
   const exportColumns = computed<ExportTableColumnItem[]>(() =>
     getExportColumns(propsRes.value.columns, customFieldsFilterConfig.value as FilterFormItem[])
